backend/requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
python-dotenv
requests
huggingface_hub
numpy
